TOK-396 Move test code from `src/pool` to `test/pool`
Also refactored `utils/pool` to make use of `TempFile`.
