+++ src/lib.rs
@@ -7,17 +7,6 @@ extern crate num_traits;
 #[macro_use]
 extern crate num_derive;
 
-#[cfg(test)]
-#[macro_use]
-extern crate serde_derive;
-
-#[cfg(test)]
-#[macro_use]
-extern crate serde_json;
-
-#[cfg(test)]
-extern crate rand;
-
 #[macro_use]
 mod macros;
