+++ src/utils/mod.rs
@@ -1,6 +1,3 @@
 pub mod results;
 pub mod callbacks;
 mod sequence;
-
-#[cfg(test)]
-pub mod test;
