--- src/utils/test/mod.rs
@@ -1,3 +1,0 @@
-pub mod file;
-pub mod pool;
-pub mod rand;
