--- src/utils/test/rand.rs
@@ -1,11 +1,0 @@
-use rand;
-use rand::distributions::{Distribution, Alphanumeric};
-
-
-pub fn random_string(n: usize) -> String{
-    let mut range = rand::thread_rng();
-    Alphanumeric
-        .sample_iter(&mut range)
-        .take(n)
-        .collect()
-}
